Create PurchaseList.java
A shopkeeper records all purchases at the end of the day. He has prices for all the purchases that happened at the shop today. He wants to know which item got sold only once and which item got sold more than once.

Write a program that reads the space-separated prices and prints the price of items that sold only once and more than once.