ChocoBoxCostCalc.java
Ram went to the store to buy chocolates. In addition to the chocolate he likes, the chocolate box also contains other chocolates he doesn't like. He told the shopkeeper he would purchase the chocolate box if it contained more than half of the chocolates he liked. 
Each chocolate price is one rupee. To make Ram purchase a chocolate box, the shopkeeper must remove the minimum number of chocolates that Ram dislikes. 
Write a program to print the final cost of the chocolate box.